// src/shared/generated/WindowsRuntimeWindowPlacementEventRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DisplayTargetRecord } from "./DisplayTargetRecord";
import type { GameWindowPlacementRecord } from "./GameWindowPlacementRecord";

export type WindowsRuntimeWindowPlacementEventRecord = { eventId: string, adapterSequence: number, nativeHostId: number, nativeGeneration: number, windowId: string, windowGeneration: number, topologyRevision: number, targetDisplay: DisplayTargetRecord, placement: GameWindowPlacementRecord, };